// brewlog_ui/src/app/bl-api/model/recipeValues.ts
/**
 * brewlog.api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { BeerStyle } from './beerStyle';


export interface RecipeValues { 
    batchName?: string | null;
    targetOg?: number;
    targetFg?: number;
    targetAbv?: number;
    targetMashPh?: number;
    mashVolume?: number;
    lauterVolume?: number;
    targetVolume?: number;
    mashTime?: number;
    boilTime?: number;
    beerStyle?: BeerStyle;
}
export namespace RecipeValues {
}